:recycle: AppProvider 기능 /utils 폴더로 이동
