+++ src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from '.';
+export * from './AppProvider';
